Update locations.json (#614)
* Update locations.json

1) Added "Port Canaveral" to correct error from Pull Request #608 by @EliW. City "Port Canaveral" was not in list.
https://travis-ci.org/szabgab/codeandtalk.com/jobs/489030530
ERROR 10: The value of city "Port Canaveral" is not in our list. If this was not a typo, add it to data/locations.json.

2) Corrected spelling of "St. Augistine" Florida to "St. Augustine"

* Update locations.json

Upon updating spelling of "St. Augustine", received ERROR 10: The value of city "St. Augistine" is not in our list.  So re-adding "St. Augistine" as well as the correct spelling.